refactor(fe): extract CameraScan
